# gateway-runtime/docs/swagger.yaml
definitions:
...
  models.RecoveryRequest:
    properties:
      after_hash:
        type: string
      agent_applied_at:
        type: string
      agent_command_id:
        type: string
      anchor_id:
        type: string
      approved_at:
        type: string
      approved_by:
        type: string
      before_hash:
        type: string
      cdc_deadline_at:
        type: string
      cdc_status:
        type: string
      client_before_hash:
        type: string
      client_id:
        type: string
      desired_state_hash:
        type: string
      executed_at:
        type: string
      executed_by:
        type: string
      execution_started_at:
        type: string
      expected_merkle_root:
        type: string
      failure_reason:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      incident_id:
        type: string
      operation:
        type: string
      reason:
        type: string
      recovery_event_id:
        type: string
      reference_anchor_id:
        type: string
      reference_log_hash:
        type: string
      reference_merkle_root:
        type: string
      requested_at:
        type: string
      requested_by:
        type: string
      resource:
        type: string
      result_audit_log_id:
        type: string
      selected_log_id:
        type: string
      snapshot_checksum:
        description: |-
          These frozen references are nullable at the database level so AutoMigrate
          can be applied safely to installations that already contain legacy
          recovery requests. New requests are rejected unless all values are set.
        type: string
      snapshot_object_key:
        type: string
      snapshot_plaintext_hash:
        type: string
      snapshot_version_id:
        type: string
      status:
        type: string
      target_log_id:
        type: string
    type: object
  recovery.CreateRequestInput:
    properties:
      idempotency_key:
        type: string
      incident_id:
        type: string
      reason:
        minLength: 5
        type: string
      selected_log_id:
        type: string
    required:
    - idempotency_key
    - incident_id
    - reason
    - selected_log_id
    type: object
  recovery.ErrorResponse:
    properties:
      error:
        example: Pesan kesalahan
        type: string
    type: object
  recovery.IncidentDetailResponse:
    properties:
      data:
        $ref: '#/definitions/recovery.IncidentDetailView'
    type: object
  recovery.IncidentDetailView:
    properties:
      client_id:
        type: string
      detected_at:
        type: string
      detected_hash:
        type: string
      detected_state_hash:
        type: string
      discrepancy_summary:
        type: string
      expected_hash:
        type: string
      id:
        type: string
      incident_scope:
        type: string
      incident_type:
        type: string
      last_confirmed_at:
        type: string
      log_id:
        type: string
      reference_log_id:
        type: string
      resolved_at:
        type: string
      resource:
        type: string
      source_status:
        type: string
      status:
        type: string
      tampered_evidence_available:
        type: boolean
      tampered_metadata: {}
    type: object
  recovery.RequestDetailResponse:
    properties:
      data:
        $ref: '#/definitions/models.RecoveryRequest'
    type: object
  recovery.RequestListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.RecoveryRequest'
        type: array
    type: object
...
info:
  contact:
    email: support@auditchain.local
    name: API Support
  description: API Enterprise untuk sistem audit log berbasis Blockchain dan Merkle
    Tree.
  termsOfService: http://swagger.io/terms/
  title: AuditChain Gateway API
  version: "1.0"
paths:
...
  /recovery/incidents/{id}:
    get:
      description: Mengambil detail sebuah insiden tamper berdasarkan ID.
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Detail Insiden Tamper
          schema:
            $ref: '#/definitions/recovery.IncidentDetailResponse'
        "401":
          description: Identitas client tidak valid
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "404":
          description: Tamper incident tidak ditemukan
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "500":
          description: Gagal mengambil detail tamper incident
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tamper incident detail
      tags:
      - Recovery
  /recovery/incidents/{id}/candidates:
    get:
      description: Mengambil referensi event client yang valid terhadap Merkle proof
        dan Fabric untuk insiden tertentu.
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Daftar kandidat recovery
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Identitas client tidak valid
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "404":
          description: Insiden tidak ditemukan
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "500":
          description: Gagal mengambil daftar kandidat
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List trusted client recovery references
      tags:
      - Recovery
  /recovery/incidents/{id}/preflight:
    post:
      description: Memeriksa referensi PostgreSQL/Fabric dan state live Agent sebelum
        recovery tanpa melakukan write.
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Hasil preflight check
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Identitas client tidak valid
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "409":
          description: Konflik atau kondisi tidak valid untuk preflight
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "500":
          description: Gagal melakukan preflight
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Preflight check for recovery
      tags:
      - Recovery
  /recovery/requests:
    get:
      description: Mengambil daftar permintaan (request) recovery yang diajukan.
      parameters:
      - description: Filter Status Request
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Daftar recovery request
          schema:
            $ref: '#/definitions/recovery.RequestListResponse'
        "401":
          description: Identitas client tidak valid
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "500":
          description: Gagal mengambil daftar recovery request
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List recovery requests
      tags:
      - Recovery
    post:
      consumes:
      - application/json
      description: Membekukan intent recovery state client melalui Agent setelah preflight
        Fabric dan state live lulus.
      parameters:
      - description: Data Request Recovery
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/recovery.CreateRequestInput'
      produces:
      - application/json
      responses:
        "201":
          description: Recovery request berhasil dibuat
          schema:
            $ref: '#/definitions/recovery.RequestDetailResponse'
        "400":
          description: Format request tidak valid
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "401":
          description: Identitas client tidak valid
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "409":
          description: Konflik atau state tidak valid
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "500":
          description: Gagal membuat recovery request
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create recovery request
      tags:
      - Recovery
  /recovery/requests/{id}/execute:
    post:
      description: Mengeksekusi intent recovery ke row client existing melalui Agent;
        tidak membuat tabel baru.
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Eksekusi recovery dimulai/berhasil
          schema:
            $ref: '#/definitions/recovery.RequestDetailResponse'
        "401":
          description: Identitas client tidak valid
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "404":
          description: Request tidak ditemukan
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "409":
          description: State request tidak sesuai
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "500":
          description: Gagal mengeksekusi recovery
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Execute recovery request
      tags:
      - Recovery
  /recovery/resources/{resource}/versions:
    get:
      description: Mengambil histori event client ter-anchor dari sebuah resource
        (table).
      parameters:
      - description: Resource/Table Name
        in: path
        name: resource
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Histori event client ter-anchor
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Identitas client tidak valid
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
        "500":
          description: Gagal mengambil histori recovery
          schema:
            $ref: '#/definitions/recovery.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List snapshot versions by resource
      tags:
      - Recovery
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"